+++ __tests__/html/withEmoji.2.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('withEmoji scenario 2', () => {
+  test('should pass', () => runHTML('withEmoji.2.html'));
+});