+++ __tests__/html/withEmoji.6.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('withEmoji scenario 6', () => {
+  test('should pass', () => runHTML('withEmoji.6.html'));
+});