+++ __tests__/html/withEmoji.8.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('withEmoji scenario 8', () => {
+  test('should pass', () => runHTML('withEmoji.8.html'));
+});