+++ __tests__/html/withEmoji.old.12.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('withEmoji old scenario 12', () => {
+  test('should pass', () => runHTML('withEmoji.old.12.html'));
+});