+++ __tests__/html/withEmoji.oneEmoji.js
@@ -1,0 +1,5 @@
+/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
+
+describe('withEmoji one emoji', () => {
+  test('should pass', () => runHTML('withEmoji.oneEmoji.html'));
+});